Django-ReactJS/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
psycopg2==2.9.6
PyJWT==2.7.0
django-cors-headers==4.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
